Refactoring Navbar + fixed mobile
