--- src/components/navbar/NavbarLink.tsx
@@ -1,15 +1,0 @@
-import { NavbarItem } from "@nextui-org/react";
-import Link from "../link-components/Link";
-
-interface NavbarLinkProps {
-  name: string;
-  href: string;
-}
-
-export default function NavbarLink({ name, href }: NavbarLinkProps) {
-  return (
-    <NavbarItem>
-      <Link name={name} href={href} />
-    </NavbarItem>
-  );
-}
