+++ src/components/navbar/menuLinks.ts
@@ -1,0 +1,15 @@
+interface menuLink {
+    name: string,
+    href: string
+}
+
+export const menuItems: menuLink[] = [
+    {
+        name: "About",
+        href: "/about"
+    },
+    {
+        name: "Games",
+        href: "/games"
+    }
+];
